RD done barebone doe for card feature

+++ app/api/card_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify, session, request
 from app.models import db, Card
-from app.forms.card_form import CardForm
+from app.forms.card_form import CardForm,EditCardForm
 from flask_login import current_user, login_user, logout_user, login_required
 
 
@@ -16,18 +16,17 @@ def get_all_user_cards():
     res = [card.to_dict() for card in user_cards]
 
     cards = Card.query.all()
-    res2 = [card.to_dict() for card in cards]
-    return {"cards": res,"all_cards":res2}
+    return {"cards": res}
 
 
 @card_routes.route("/new",methods=["POST"])
 @login_required
 def post_new_card():
     '''
-    Posts a new Cardnk
+    Posts a new Card
     - Grabs the information sent from the user
     - Passes that information through wtforms validators
-    - If they are validated on submit then submits a new credit card for the currenkt user
+    - If they are validated on submit then submits a new credit card for the curret user
     - If there are any validation errors, then they will be sent back to the front end
     '''
 
@@ -54,7 +53,54 @@ def post_new_card():
         print("There were some form errors", form.errors)
         return {"errors": form.errors}, 400, {"Content-Type": "application/json"}
 
+@card_routes.route("/<int:id>")
+@login_required
+def get_single_card(id):
+    single_card = Card.query.get(id)
 
+    if single_card.owner_id != current_user.id:
+        return {"error":"unauthorized"}
+    if single_card is None:
+        return {"error":"Card can not be found"}
+    
+    return {"card": single_card.to_dict()}
+
+@card_routes.route("/<int:id>",methods=["PUT"])
+@login_required
+def update_card(id):
+    '''
+    User sends an ID to the backend
+    Backend grabs that id and does some validation(Does the user own this card or is it validated card)
+    If Updated Card Passes through all the validations then we will update that card.
+    Otherwise the backend will send errors that the user will need to fix
+    '''
+
+    updated_card = Card.query.get(id)
+
+    if updated_card.owner_id != current_user.id:
+        return {"errors": "Unauthorized"},401
+    
+    if updated_card is None:
+        return {"errors": "Card Could not be found"},404
+
+    form = EditCardForm()
+    form["csrf_token"].data = request.cookies["csrf_token"]
+
+    if form.validate_on_submit():
+        data = form.data
+        if data["card_number"]:
+            updated_card.card_number = data["card_number"]
+        if data["bank_name"]:
+            updated_card.bank_name = data["bank_name"]
+        
+        db.session.commit()
+        return {"card": updated_card.to_dict()}
+    if form.errors:
+        print("There were some errors")
+        return{"errors":form.errors},400,
+
+
+    pass
 
 @card_routes.route("/<int:id>",methods=["DELETE"])
 @login_required
@@ -69,6 +115,9 @@ def delete_card(id):
 
     if(current_user.id != deleted_card.owner_id):
         return {"errors": "unauthorized"},400
+    
+    if deleted_card is None:
+        return {"errors": "Card could not be found"},404
 
     db.session.delete(deleted_card)
     db.session.commit()
